Final README for REU
This is a smaller version of the readme. The bigger updates to the readme will reside on the final implementation of the tri modal model. Please feel free to browse the model folder to view my work and to my peers always feel free to reach out if you have questions. Thank you for working with me.